Working on jump shotting
Made custom velocity tracker script for the controllers to give to the ball when throwing. Ball will use this velocity instead of the interactor script's calculations since there is an issue with the ball shooting off in a different direction at really high speeds when trying to shoot while jumping.